requirements.txt
blinker==1.8.2
click==8.1.8
flask==3.0.3
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
MarkupSafe==2.1.5
mysql-connector==2.2.9
werkzeug==3.0.6
zipp==3.20.2
